Retry triggering builds if they fail
We hit a buildkite outage today and it took down the service.  Instead
of dying, pause and retry.

A fixed 30 second retry isn't the right answer.  We should be doing some
sort of exponential backoff.

Also, this blocks handling any more SSH events.  Probably fine for now,
but we can revisit that if this becomes an issue with a goroutine.
Hopefully buildkite won't be down for a while.